Moved and added convert_binary_to_2d
