+++ golpi/__init__.py
@@ -1,1 +1,2 @@
 from .golpi import *
+from .utility import *
